src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class StatsServiceTest {
    @Test
    void shouldReturnSum() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;
        int actual = service.shouldReturnSum(sales);
        assertEquals(expected, actual);

    }

    @Test
    void shouldReturnAverage() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;
        int actual = service.shouldReturnAverage(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldReturnMontWithMaxSales() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 8;
        int actual = service.shouldReturnMontWithMaxSales(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldReturnMontWithMinSales() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 9;
        int actual = service.shouldReturnMontWithMinSales(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldReturnLowerThanAverage() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.shouldReturnLowerThanAverage(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldReturnHigherThanAverage() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.shouldReturnLowerThanAverage(sales);
        assertEquals(expected, actual);
    }
}
